Refactoring.
Renamed some inner classes and local variables. The classes were renamed to reduce naming redundancy. For example, the previously called "InstructionType" class has been renamed to simply "Type", since it is an inner class of "Instruction".

Also as part of this commit, the imports in all classes have been re-optimized with a slightly different import order.